added documentation

+++ ASSIGNMENT_04/controllers/productController.js
@@ -1,69 +1,68 @@
-const productModel = require("../models/productModel");
+const productModel = require("../models/productModel");//getting a productModel i.e a schema of product
 
 const showAddProductPage=(req,res)=>{
     res.render("addproduct")
-}
+};//rendering a add product page
 
+//implementing a showUpdateProduct basically a update product page
 const showUpdateProduct=async(req,res)=>{
-    const {id}=req.params;
-    console.log(id);
+    const {id}=req.params;//getting id through params 
     try {
-
-        let data=await productModel.findOne({_id:id});
-        console.log(data);
-        return res.render("updateproduct",{data:data})
+        let data=await productModel.findOne({_id:id});//finding a data(document) oject using object id->userid
+        return res.render("updateproduct",{data:data});//rendering update product and passing document data
         
     } catch (error) {
-        
+        console.log(error);//to handle error if promise reject
     }
-    res.render("updateproduct")
-}
+    res.render("updateproduct");//rednering a update product form/page
+};
 
+// implementing a saveproduct function using async await
 const saveProduct=async(req,res)=>{
 
-    const bodyData=req.body;
-    const product=new productModel(bodyData);
+    const bodyData=req.body;//getting form field in oject format
+    const product=new productModel(bodyData);//creating a new instance of product model and pass data in schema
 
     try {
-        await product.save();
-        res.redirect("/")
+        await product.save();//saving data in DB
+        res.redirect("/");//after saving data redirecting to home
     } catch (error) {
-        console.log(error);
+        console.log(error);//to catch a error
     }
+};
 
-}
-
+// implementing a delete product function using async-await
 const deleteProduct=async(req,res)=>{
 
-    const {id}=req.params;
+    const {id}=req.params;//getting id through params 
     try {
-        await productModel.findByIdAndRemove({_id:id})
+        await productModel.findByIdAndRemove({_id:id})//finding and removing document using userId
         res.redirect("/")
     } catch (error) {
-        console.log(error);
+        console.log(error);//error handling
     }    
-}
+};
 
+// implementing a getAllproducts function using async-await
 const getAllProducts=async(req,res)=>{
     try {
-        let products = await productModel.find({});
-        console.log(products);
-        res.render("index",{products:products})
+        let products = await productModel.find({}); //find all documents and saving in products variable
+        res.render("index",{products:products});//rednering homepage and passing products data
     } catch (error) {
-        console.log(error);
+        console.log(error);//error handling
     }
-}
+};
 
+// implementing a updateProduct function using async-await
 const updateProduct=async(req,res)=>{
     try {
-
-        const bodyData=req.body;
-        const {id}=req.params;
-        await productModel.findByIdAndUpdate({_id:id},{$set:bodyData})
-        res.redirect("/")
+        const bodyData=req.body;//getting body data in object format
+        const {id}=req.params;//getting id through params
+        await productModel.findByIdAndUpdate({_id:id},{$set:bodyData});//updating data using $set query
+        res.redirect("/"); //once updated redirect to homepage
         
     } catch (error) {
-        console.log(error);
+        console.log(error);//error handling
     }
 }
 
@@ -74,4 +73,4 @@ module.exports={
     saveProduct,
     showAddProductPage,
     showUpdateProduct
-}
+};//exporting all the functions
